+++ packages/components/src/components/banner.stories.tsx
@@ -29,7 +29,8 @@ export const Default = Template.bind({});
 Default.args = {
   children: (
     <>
-      <b>Important:</b> This documentation covers GraphQL Config v3. For the 2.x doc, check: graphql-config.com/legacy.
+      <b>Important:</b> This documentation covers GraphQL Config v3. For the 2.x doc, check:
+      graphql-config.com/legacy.
     </>
   ),
 };