+++ packages/components/src/components/feature-list.tsx
@@ -4,7 +4,13 @@ import { IFeatureListProps } from '../types/components';
 import { Anchor } from './anchor';
 import { Image } from './image';
 
-export const FeatureList = ({ title, description, items, link, className }: IFeatureListProps): ReactElement => (
+export const FeatureList = ({
+  title,
+  description,
+  items,
+  link,
+  className,
+}: IFeatureListProps): ReactElement => (
   <section className={clsx('bg-white dark:bg-[#111]', className)}>
     <div className="container py-14">
       {title && (
@@ -13,10 +19,15 @@ export const FeatureList = ({ title, description, items, link, className }: IFea
             {title}
           </h2>
           {description && (
-            <div className="mt-1 mb-1.5 text-center text-base text-gray-500 dark:text-gray-400">{description}</div>
+            <div className="mt-1 mb-1.5 text-center text-base text-gray-500 dark:text-gray-400">
+              {description}
+            </div>
           )}
           {link && (
-            <Anchor {...link} className={clsx('w-max text-sm text-cyan-400 hover:text-cyan-300', link.className)} />
+            <Anchor
+              {...link}
+              className={clsx('w-max text-sm text-cyan-400 hover:text-cyan-300', link.className)}
+            />
           )}
         </div>
       )}
@@ -25,11 +36,16 @@ export const FeatureList = ({ title, description, items, link, className }: IFea
           <div className="flex w-56 flex-col items-center" key={item.title}>
             <Image {...item.image} className={clsx('w-28', item.image.className)} />
             <h3 className="m-0 text-lg font-bold text-black dark:text-gray-50">{item.title}</h3>
-            <div className="text-center text-sm text-gray-500 dark:text-gray-400">{item.description}</div>
+            <div className="text-center text-sm text-gray-500 dark:text-gray-400">
+              {item.description}
+            </div>
             {item.link && (
               <Anchor
                 {...item.link}
-                className={clsx('mt-2 w-max text-sm text-cyan-400 hover:text-cyan-300', item.link.className)}
+                className={clsx(
+                  'mt-2 w-max text-sm text-cyan-400 hover:text-cyan-300',
+                  item.link.className,
+                )}
               />
             )}
           </div>