+++ packages/components/src/components/search-bar-v2/side-preview.tsx
@@ -19,12 +19,17 @@ export const SidePreview = ({
   }
   const project = products.find(p => p.name === item.source);
   return (
-    <div className="flex h-[600px] select-text flex-col overflow-y-auto p-10" style={{ '--accentColor': accentColor }}>
+    <div
+      className="flex h-[600px] select-text flex-col overflow-y-auto p-10"
+      style={{ '--accentColor': accentColor }}
+    >
       {project && (
         <div className="my-1 mb-4 box-border flex w-full select-none items-center justify-center">
           <project.logo className="mr-3 h-9 w-9" />
           <span className="flex flex-col justify-center">
-            <h4 className="m-0 text-sm font-semibold text-black dark:text-gray-400">{project.name}</h4>
+            <h4 className="m-0 text-sm font-semibold text-black dark:text-gray-400">
+              {project.name}
+            </h4>
           </span>
         </div>
       )}
@@ -53,7 +58,9 @@ export const SidePreview = ({
       </p>
       {item.toc.length > 1 && (
         <>
-          <h4 className="mt-10 mb-1 w-full text-lg font-light text-gray-600 dark:text-gray-400">Table of content:</h4>
+          <h4 className="mt-10 mb-1 w-full text-lg font-light text-gray-600 dark:text-gray-400">
+            Table of content:
+          </h4>
           <ol className="text-left text-base font-light text-gray-600 dark:text-gray-400">
             {item.toc.map(tocItem => (
               <li key={tocItem.title}>{tocItem.title}</li>