+++ packages/components/src/components/solutions-menu.tsx
@@ -65,18 +65,24 @@ const CATEGORIES: {
   },
 ];
 
-export const SolutionsMenu = forwardRef<HTMLDivElement>(function SolutionsMenu(_, forwardedRef): ReactElement {
+export const SolutionsMenu = forwardRef<HTMLDivElement>(function SolutionsMenu(
+  _,
+  forwardedRef,
+): ReactElement {
   return (
     <div
       className="w-[700px] rounded-lg bg-white p-5 dark:bg-neutral-800"
       style={{
-        boxShadow: 'hsl(206 22% 7% / 35%) 0 10px 38px -10px, hsl(206 22% 7% / 20%) 0 10px 20px -15px',
+        boxShadow:
+          'hsl(206 22% 7% / 35%) 0 10px 38px -10px, hsl(206 22% 7% / 20%) 0 10px 20px -15px',
       }}
       ref={forwardedRef}
     >
       {CATEGORIES.map(category => (
         <div key={category.title} className="mb-6 last:mb-0">
-          <h3 className="mt-1 mb-5 w-full text-base font-normal text-black dark:text-gray-400">{category.title}</h3>
+          <h3 className="mt-1 mb-5 w-full text-base font-normal text-black dark:text-gray-400">
+            {category.title}
+          </h3>
           <div className="flex flex-wrap">
             {category.items.map(item => (
               <Anchor