+++ packages/components/src/env.d.ts
@@ -1,6 +1,8 @@
 declare module '@next/bundle-analyzer' {
   import { NextConfig } from 'next';
-  export default function nextBundleAnalyzer(config: { enabled: boolean }): (nextConfig: NextConfig) => NextConfig;
+  export default function nextBundleAnalyzer(config: {
+    enabled: boolean;
+  }): (nextConfig: NextConfig) => NextConfig;
 }
 
 declare module 'next-videos' {