+++ packages/algolia/CHANGELOG.md
@@ -4,47 +4,68 @@
 
 ### Patch Changes
 
-- [#482](https://github.com/the-guild-org/the-guild-components/pull/482) [`7e84a97`](https://github.com/the-guild-org/the-guild-components/commit/7e84a9799816f895e69d1facec0e88ea7a8467d5) Thanks [@charlypoly](https://github.com/charlypoly)! - dependencies updates:
+- [#482](https://github.com/the-guild-org/the-guild-components/pull/482)
+  [`7e84a97`](https://github.com/the-guild-org/the-guild-components/commit/7e84a9799816f895e69d1facec0e88ea7a8467d5)
+  Thanks [@charlypoly](https://github.com/charlypoly)! - dependencies updates:
 
-  - Added dependency [`commander@^9.4.1` ↗︎](https://www.npmjs.com/package/commander/v/9.4.1) (to `dependencies`)
+  - Added dependency [`commander@^9.4.1` ↗︎](https://www.npmjs.com/package/commander/v/9.4.1) (to
+    `dependencies`)
 
-- [#482](https://github.com/the-guild-org/the-guild-components/pull/482) [`7e84a97`](https://github.com/the-guild-org/the-guild-components/commit/7e84a9799816f895e69d1facec0e88ea7a8467d5) Thanks [@charlypoly](https://github.com/charlypoly)! - support from node 12
+- [#482](https://github.com/the-guild-org/the-guild-components/pull/482)
+  [`7e84a97`](https://github.com/the-guild-org/the-guild-components/commit/7e84a9799816f895e69d1facec0e88ea7a8467d5)
+  Thanks [@charlypoly](https://github.com/charlypoly)! - support from node 12
 
 ## 1.0.0
 
 ### Major Changes
 
-- [#469](https://github.com/the-guild-org/the-guild-components/pull/469) [`21bf61e`](https://github.com/the-guild-org/the-guild-components/commit/21bf61e6f91507066a462d8ec9383ea4a0efd9e0) Thanks [@charlypoly](https://github.com/charlypoly)! - Algolia scripts as binary for shared CI worflows
+- [#469](https://github.com/the-guild-org/the-guild-components/pull/469)
+  [`21bf61e`](https://github.com/the-guild-org/the-guild-components/commit/21bf61e6f91507066a462d8ec9383ea4a0efd9e0)
+  Thanks [@charlypoly](https://github.com/charlypoly)! - Algolia scripts as binary for shared CI
+  worflows
 
 ## 0.1.0
 
 ### Minor Changes
 
-- [#445](https://github.com/the-guild-org/the-guild-components/pull/445) [`a34d3a6`](https://github.com/the-guild-org/the-guild-components/commit/a34d3a65c197639bdb2fa9f10c9e7866562851de) Thanks [@B2o5T](https://github.com/B2o5T)! - use tsup for bundling
+- [#445](https://github.com/the-guild-org/the-guild-components/pull/445)
+  [`a34d3a6`](https://github.com/the-guild-org/the-guild-components/commit/a34d3a65c197639bdb2fa9f10c9e7866562851de)
+  Thanks [@B2o5T](https://github.com/B2o5T)! - use tsup for bundling
 
 ## 0.0.2
 
 ### Patch Changes
 
-- [#437](https://github.com/the-guild-org/the-guild-components/pull/437) [`0226558`](https://github.com/the-guild-org/the-guild-components/commit/0226558fc38fa3952a10dc3e4b531fce10e742df) Thanks [@B2o5T](https://github.com/B2o5T)! - fix algolia, replace `meta.json` by `_meta.json`
+- [#437](https://github.com/the-guild-org/the-guild-components/pull/437)
+  [`0226558`](https://github.com/the-guild-org/the-guild-components/commit/0226558fc38fa3952a10dc3e4b531fce10e742df)
+  Thanks [@B2o5T](https://github.com/B2o5T)! - fix algolia, replace `meta.json` by `_meta.json`
 
 ## 0.0.1
 
 ### Patch Changes
 
-- [#423](https://github.com/the-guild-org/the-guild-components/pull/423) [`87f9420`](https://github.com/the-guild-org/the-guild-components/commit/87f942002cb37b6f082109e33b81d5e591b1a632) Thanks [@B2o5T](https://github.com/B2o5T)! - bump nextra, move `guild-docs` to `@theguild/components` and `@theguild/algolia` to `guild-components` monorepo
+- [#423](https://github.com/the-guild-org/the-guild-components/pull/423)
+  [`87f9420`](https://github.com/the-guild-org/the-guild-components/commit/87f942002cb37b6f082109e33b81d5e591b1a632)
+  Thanks [@B2o5T](https://github.com/B2o5T)! - bump nextra, move `guild-docs` to
+  `@theguild/components` and `@theguild/algolia` to `guild-components` monorepo
 
 ## 0.2.2
 
 ### Patch Changes
 
-- [#660](https://github.com/the-guild-org/the-guild-docs/pull/660) [`5dbc5c1`](https://github.com/the-guild-org/the-guild-docs/commit/5dbc5c1c4002e52056f6757e42538788d08e10b1) Thanks [@charlypoly](https://github.com/charlypoly)! - feat(algolia): hash content for lockfile to trigger reindexing on content change
+- [#660](https://github.com/the-guild-org/the-guild-docs/pull/660)
+  [`5dbc5c1`](https://github.com/the-guild-org/the-guild-docs/commit/5dbc5c1c4002e52056f6757e42538788d08e10b1)
+  Thanks [@charlypoly](https://github.com/charlypoly)! - feat(algolia): hash content for lockfile to
+  trigger reindexing on content change
 
 ## 0.2.1
 
 ### Patch Changes
 
-- [#658](https://github.com/the-guild-org/the-guild-docs/pull/658) [`287e657`](https://github.com/the-guild-org/the-guild-docs/commit/287e657881fc2744c457700b002f7a979b2b0b5c) Thanks [@charlypoly](https://github.com/charlypoly)! - fix(algolia): remove `import` and `export` from records content
+- [#658](https://github.com/the-guild-org/the-guild-docs/pull/658)
+  [`287e657`](https://github.com/the-guild-org/the-guild-docs/commit/287e657881fc2744c457700b002f7a979b2b0b5c)
+  Thanks [@charlypoly](https://github.com/charlypoly)! - fix(algolia): remove `import` and `export`
+  from records content
 
 ## 0.2.0